# [Xamarin.Android.Build.Tasks] don't use so many temp files