@jonathanpeppers 2018-12-07
Actually looks like I could use a `TextWriter` overload here: https://github.com/xamarin/xamarin-android/blob/24517ef0095dbcb8bdd2e88edb41227215c96b37/src/Xamarin.Android.Build.Tasks/Generator/Generator.cs#L49

So I could remove *all* the temp files in `GenerateJavaStubs`, without a Java.Interop change... I'll try it.