@jonathanpeppers 2018-12-07
> So I could remove all the temp files in GenerateJavaStubs, without a Java.Interop change... I'll try it.

Nope: https://github.com/xamarin/xamarin-android/commit/d510ca0a587c30aeb596083bf46867d1b05634f3

Will need to add one public method here in Java.Interop, something like:
```csharp
public string GetDestinationPath (string outputPath)
{
	string path = outputPath;
	foreach (string dir in package.Split ('.'))
		path = Path.Combine (path, dir);

	if (!Directory.Exists (path))
		Directory.CreateDirectory (path);

	path = Path.Combine (path, name + ".java");
	return path;
}
```
`package` is private to `JavaCallableWrapperGenerator`.

So we can take this PR as a first step, and do more later.